Rename entityMapping to entityLinking
